Rationalise value formatting and fix tests
re #576

+++ src/js/test/spec/models/visualisation/element/dimensions.spec.js
@@ -9,7 +9,7 @@ define(['models/dataset', 'models/dataset/connection', 'models/visualisation/ele
                     visualisation_id: 'test02',
                     fields: [
                         {id: 'test04', type: 'string'},
-                        {id: 'test05', type: 'numeric'}
+                        {id: 'test05', type: 'integer'}
                     ]
                 });
             this.dataset.reset();
