+++ src/js/test/spec/views/element/d3/line.spec.js
@@ -13,7 +13,7 @@ define(['jquery', 'models/dataset', 'models/dataset/connection', 'models/visuali
                     visualisation_id: 'test02',
                     fields: [
                         {id: 'test04', type: 'string'},
-                        {id: 'test05', type: 'numeric'}
+                        {id: 'test05', type: 'integer'}
                     ]
                 });
             this.dataset.reset();
